// src/theme/theme.interfaces.ts
export interface ITheme {
  theme: {
    colors: {
      black: string;
      white: string;
      green: string;
      red: string;
      blue: string;
      progress_audio_color: string;
      bright_blue: string;
      typing_bubble_color: string;
      background_dialogItem_color: string;
      bubble_color: string;
      background_color: string;
      bright_grey: string;
      gray: string;
      dark_grey: string;
      purple: string;
      background_modal_color: string;
    };
  };
}